Separate handshake logic, and disable cookie auth in browsers
Cookie auth is intended to demonstrate that you can access a local user
private file (the cookie) to show you are authenticated as that user.
Clearly in a browser we can't do this - so we swap in a separate
implementation that skips this entirely. Conveniently this also lets us
drop more node deps - removing the dependency on fs, path & crypto.