Drop recursive introspection workaround
This appears to be a workaround for interfaces that expose introspection
weirdly. Doesn't affect Frida, and it significantly complicates some
changes coming up here, so I'm going to drop this.

Might affect PulseAudio D-Bus communication? If you run into issues and
you find this commit, please open an issue on this repo and we can bring
this workaround back.